pyproject.toml
[project]
name = "cot-loop"
version = "0.1.0"
description = "Repro scripts for looping analysis (Figure 1)."
readme = "FIG1_REPRO.md"
requires-python = ">=3.10"

dependencies = [
  "datasets>=2.16.0",
  "matplotlib>=3.7.0",
  "transformers==4.57.3",
  "math-verify[antlr4-13-2]>=0.9.0",
  "vllm>=0.15.0 ; sys_platform != 'darwin'",
  "wandb>=0.17.0",
  "python-dotenv>=1.0.0",
  "scikit-learn>=1.4.0",
  "sentencepiece>=0.1.99",
  "tiktoken>=0.5.2",
  "protobuf>=4.25.0",
  "torch==2.9.1",
  "flash_attn==2.8.3 ; sys_platform != 'darwin'",
  "ninja>=1.13.0",
  "accelerate>=1.12.0",
]

[tool.uv]
package = false

[tool.uv.extra-build-dependencies]
flash-attn = ["torch==2.9.1"]
